+++ tests/test_flickr_url_parser.py
@@ -58,6 +58,7 @@ def test_it_rejects_a_url_which_isnt_flickr(url: str) -> None:
         "https://c8.staticflickr.com/6/5159/142abc431_7cf094b085_b.jpg",
         "farm3.static.flickr.com",
         "https://www.flickr.com/photo.gne?short=-1",
+        "https://www.flickr.com/apps/video/stewart.swf?photo_id=-1",
     ],
 )
 def test_it_rejects_a_flickr_url_which_does_not_have_photos(url: str) -> None:
@@ -179,6 +180,14 @@ def test_it_can_parse_the_homepage(url: str) -> None:
         (
             "https://live.staticflickr.com/video/52868534222/346a41e5a9/1080p.mp4",
             "52868534222",
+        ),
+        #
+        # This URL comes from the flickr.photos.getSizes API for
+        # this photo.
+        # Retrieved 2 January 2024
+        (
+            "https://www.flickr.com/apps/video/stewart.swf?v=2968162862&photo_id=53262935176&photo_secret=06c382eee3",
+            "53262935176",
         ),
     ],
 )